fix run
