+++ .github/workflows/wpt.yml
@@ -135,6 +135,8 @@ jobs:
           cache: 'pip'
       - name: Set up virtualenv
         run: pip install virtualenv
+      - name: Install pipenv
+        run: pip install pipenv
       - name: Install python dependencies
         run: pipenv install
       - name: Set up hosts
