+++ tools/wpt-report-builder/builder.mjs
@@ -17,7 +17,7 @@
 import child_process from 'child_process';
 import fs from 'fs';
 
-import {generateReport} from './formatter.mjs';
+import {escapeHtml, generateReport} from './formatter.mjs';
 
 function getCurrentCommit() {
   if (process.env.GITHUB_SHA) {
@@ -40,7 +40,39 @@ function getOutputPath() {
   return process.argv.slice(2)[1];
 }
 
+function getAllTests() {
+  const rawCommandResult = child_process
+    .execSync(
+      // Magic command line that makes required pytest imports and gets the list of
+      // tests. Details: go/webdriver:wpt-total-test-count.
+      '(cd ./wpt/webdriver/tests/bidi; PYTHONPATH="$( pwd )/../../../tools/webdriver:$( pwd )/../../../tools/third_party/websockets/src:$( pwd ):$( pwd )../../../tools/webdriver/webdriver/bidi/modules/permissions.py:$( pwd )/../../.."  pytest --collect-only --rootdir=../../.. -o=\'python_files=*.py\' --quiet)'
+    )
+    .toString();
+
+  const tests = [];
+  rawCommandResult.split('\n').forEach((line) => {
+    if (line.startsWith('webdriver/tests/bidi')) {
+      const [testPath, ...testNameParts] = line.split('::');
+      const testName = testNameParts.join('::');
+      tests.push({
+        path: `/${testPath}/${escapeHtml(testName)}`,
+        name: testName,
+        status: 'SKIPPED',
+        message: null,
+      });
+    }
+  });
+  return tests;
+}
+
+if (process.argv.slice(2).length !== 2) {
+  console.error(
+    'Should be run from the root folder of the project.\nUsage: node tools/wpt-report-builder/builder.mjs <reportPath> <outputPath>'
+  );
+  process.exit(1);
+}
+
 fs.writeFileSync(
   getOutputPath(),
-  generateReport(readReport(getReportPath()), getCurrentCommit())
+  generateReport(readReport(getReportPath()), getAllTests(), getCurrentCommit())
 );
